CommentAnaysisExtension/backend/requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0

# ML/AI
mlflow==2.8.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
nltk==3.8.1
dagshub

# HTTP/Networking
requests==2.31.0

# Security
cryptography==41.0.7

# Monitoring & Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Code Quality
flake8==6.1.0
black==23.12.1
mypy==1.7.1

# Utilities
python-dotenv==1.0.0